# Remove the requested_images and requested_glyphs hash sets.