--- a/webrender/src/frame.rs
+++ b/webrender/src/frame.rs
@@ -18,7 +18,6 @@ use mask_cache::ClipRegion;
 use profiler::{GpuCacheProfileCounters, TextureCacheProfileCounters};
 use resource_cache::{ResourceCache, TiledImageMap};
 use scene::{Scene, SceneProperties};
-use std::cmp;
 use tiling::{CompositeOps, DisplayListMap, PrimitiveFlags};
 use util::{ComplexClipRegionHelpers, subtract_rect};
 
@@ -1166,10 +1165,6 @@ impl Frame {
                                      device_pixel_ratio,
                                      texture_cache_profile,
                                      gpu_cache_profile);
-        // Expire any resources that haven't been used for `cache_expiry_frames`.
-        let num_frames_back = self.frame_builder_config.cache_expiry_frames;
-        let expiry_frame = FrameId(cmp::max(num_frames_back, self.id.0) - num_frames_back);
-        resource_cache.expire_old_resources(expiry_frame);
         frame
     }