@glennw 2017-08-02
I think this is the last of the obvious wins in the resource cache hashing code. I have some ideas for more improvements but the hashing no longer shows up near the top of the profile with this and the previous changes...