@bors-servo 2017-08-04
:pushpin: Commit 99043f4 has been approved by `kvark`

<!-- @bors-servo r=kvark 99043f4755f5f755c3860478ec4de920baa7d353 -->