@kvark 2017-08-05
Oh fun:
> c++: internal compiler error: Killed (program cc1plus)

ICC when building Angle