@bors-servo 2017-08-06
:sunny: Test successful - [status-travis](https://travis-ci.org/servo/webrender/builds/261624351?utm_source=github_status&utm_medium=notification)
Approved by: kvark
Pushing 1b223cfa3be2f68ea8ba25290ebf0900a9622e49 to master...